requirements.txt
agency-swarm>=0.4.4
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
openai>=1.3.0
gradio>=4.19.0 